requirements.txt
agora_token_builder==1.0.0
aioredis==2.0.1
aiosqlite==0.17.0
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.0
autopep8==2.3.1
babel==2.16.0
bbcode==1.1.0
bcrypt==4.2.0
billiard==4.2.0
bleach==6.1.0
boto3==1.35.0
botocore==1.35.0
bump2version==1.0.1
bumpversion==0.6.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
codecov==2.1.13
colorama==0.4.6
constantly==23.10.4
coverage==7.6.1
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.8
django-admin-interface==0.28.8
django-allauth==64.2.1
django-appconf==1.0.6
django-braces==1.15.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cleanup==8.1.0
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-haystack==3.3.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mailer==2.3.2
django-model-utils==4.5.1
django-mptt==0.16.0
django-mysql==4.14.0
django-paypal==2.1
django-private-chat2==1.0.2
django-pure-pagination==0.3.0
django-sslserver==0.22
django-summernote==0.8.20.0
django-taggit==6.0.0
django-test-plus==2.2.4
django-timezone-field==7.0
django-tinymce==4.1.0
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
docutils==0.20.1
elastic-transport==8.15.0
elasticsearch==8.15.1
exceptiongroup==1.2.2
expiringdict==1.2.2
fabric==3.2.2
Faker==28.4.1
filelock==3.15.4
flake8==7.1.1
future==1.0.0
gitdb==4.0.11
gitdb2==4.0.2
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
html5lib==1.1
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
imageio==2.35.0
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib_metadata==8.4.0
incremental==24.7.2
invoke==2.2.0
iso8601==1.1.0
Jinja2==3.1.4
jmespath==1.0.1
jsmin==3.0.1
kombu==5.4.0
lockfile==0.12.2
lxml==5.3.0
Markdown==3.7
markdown2==2.5.0
MarkupSafe==2.1.5
mccabe==0.7.0
mistletoe==1.4.0
mock==5.1.0
model-bakery==1.19.5
moviepy==0.2.3.1
msgpack==1.0.8
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
paramiko==3.5.0
pep8==1.7.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
proglog==0.1.10
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.2.0
Pygments==2.18.0
PyHamcrest==2.1.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pypandoc==1.13
pyparsing==3.1.4
pypdf==4.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pypika-tortoise==0.1.6
pypng==0.20220715.0
pyproject-api==1.7.1
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
sendgrid==6.11.0
service-identity==24.1.0
setuptools==72.2.0
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
smmap2==3.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
stripe==10.8.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tortoise-orm==0.21.6
tox==4.18.0
tqdm==4.66.5
Twisted==24.7.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
virtualenv==20.26.3
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.4
wheel==0.44.0
whitenoise==6.7.0
Whoosh==2.7.4
wrapt==1.16.0
xhtml2pdf==0.2.16
zipp==3.20.1
zope.interface==7.0.1
